0.320 - removed infinite pager from equalizer - adjusted edit track jaudio tagger huawei bug - added lint clean architecture imports check

+++ app/src/main/java/dev/olog/msc/presentation/equalizer/EqHelper.kt
@@ -7,7 +7,7 @@ object EqHelper {
 
     fun projectY(dB: Float): Float {
         val pos = (dB - minDB) / (maxDB - minDB)
-        return (1 - pos).toFloat()
+        return (1 - pos)
     }
 
 }
